feat: 파일 첨부하여 게시글 등록하기 기능구현

+++ src/main/java/org/gosang/domain/BoardAttachVO.java
@@ -10,5 +10,5 @@ public class BoardAttachVO {
 	private String fileName;
 	private boolean fileType;
 	
-	private Long bno;
+	private int bno;
 }
